requirements.txt
Django==1.6.5
djangorestframework==2.3.14

mock==1.0.1
nose==1.3.3
django_nose==1.2
selenium==2.42.1